增加

+++ src/main/java/com/furui/domain/RealTimeResult.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -33,7 +34,7 @@ public class RealTimeResult implements Serializable {
     /**
      * 累计收入
      */
-    private double revenue;
+    private BigDecimal revenue;
 
 
 }
